requirements.txt
flask
openai-whisper
gtts
googletrans==4.0.0-rc1
